fix linux tests
