+++ Tests/Transport/StreamBufferTests.swift
@@ -68,3 +68,55 @@ class StreamBufferTests: XCTestCase {
         XCTAssert(testStream.timeout == 42, "stream buffer should set underlying timeout")
     }
 }
+
+
+final class TestStream: Transport.Stream {
+    var closed: Bool
+    var buffer: Bytes
+    var timeout: Double = -1
+    // number of times flush was called
+    var flushedCount = 0
+
+    func setTimeout(_ timeout: Double) throws {
+        self.timeout = timeout
+    }
+
+    init() {
+        closed = false
+        buffer = []
+    }
+
+    func close() throws {
+        if !closed {
+            closed = true
+        }
+    }
+
+    func send(_ bytes: Bytes) throws {
+        closed = false
+        buffer += bytes
+    }
+
+    func flush() throws {
+        flushedCount += 1
+    }
+
+    func receive(max: Int) throws -> Bytes {
+        if buffer.count == 0 {
+            try close()
+            return []
+        }
+
+        if max >= buffer.count {
+            try close()
+            let data = buffer
+            buffer = []
+            return data
+        }
+
+        let data = buffer[0..<max]
+        buffer.removeFirst(max)
+        
+        return Bytes(data)
+    }
+}
